Refactor utility widgets
