+++ src/client/kotlin/liltojustice/trueadaptivemusic/client/gui/widget/utility/DropdownWidget.kt
@@ -1,4 +1,4 @@
-package liltojustice.trueadaptivemusic.client.gui.widget
+package liltojustice.trueadaptivemusic.client.gui.widget.utility
 
 import net.minecraft.client.gui.DrawContext
 import net.minecraft.client.gui.screen.narration.NarrationMessageBuilder
@@ -36,7 +36,8 @@ class DropdownWidget(
         0,
         realizedWidth,
         textRenderer.fontHeight + TEXT_HEIGHT_BUFFER,
-        Text.literal("Dropdown Search"))
+        Text.literal("Dropdown Search")
+    )
     private val selectedOptionWidget = ClickableTextWidget(
         notSelectedPlaceholder ?: startingOption.ifEmpty { null } ?: options.firstOrNull() ?: "",
         onClick = { screen?.focused = textInputWidget },
