+++ src/client/kotlin/liltojustice/trueadaptivemusic/client/gui/widget/utility/MultiSelectDropdownWidget.kt
@@ -1,4 +1,4 @@
-package liltojustice.trueadaptivemusic.client.gui.widget
+package liltojustice.trueadaptivemusic.client.gui.widget.utility
 
 import net.minecraft.client.gui.DrawContext
 import net.minecraft.client.gui.screen.narration.NarrationMessageBuilder
@@ -48,7 +48,8 @@ class MultiSelectDropdownWidget(
                     "",
                     onHoverOption,
                     x,
-                    y)
+                    y
+                )
             },
             "dropdown"
         ) as DropdownWidget
